Drop `stateful` distinction in favor of single interpreter
This is a move only change and does not introduce any implementation differences. The code that was previously in `interpret::stateful` has been moved to `interpret` so that the stateful distinction can be dropped in favor of what was already the single interpreter implementation.
