+++ compiler/qsc/benches/eval.rs
@@ -3,7 +3,7 @@
 
 use criterion::{criterion_group, criterion_main, Criterion};
 use indoc::indoc;
-use qsc::{interpret::stateful, PackageType};
+use qsc::{interpret::Interpreter, PackageType};
 use qsc_eval::output::GenericReceiver;
 use qsc_frontend::compile::{RuntimeCapabilityFlags, SourceMap};
 
@@ -14,7 +14,7 @@ const LARGE: &str = include_str!("./large.qs");
 pub fn teleport(c: &mut Criterion) {
     c.bench_function("Teleport evaluation", |b| {
         let sources = SourceMap::new([("Teleportation.qs".into(), TELEPORT.into())], None);
-        let mut evaluator = stateful::Interpreter::new(
+        let mut evaluator = Interpreter::new(
             true,
             sources,
             PackageType::Exe,
@@ -32,7 +32,7 @@ pub fn teleport(c: &mut Criterion) {
 pub fn deutsch_jozsa(c: &mut Criterion) {
     c.bench_function("Deutsch-Jozsa evaluation", |b| {
         let sources = SourceMap::new([("DeutschJozsa.qs".into(), DEUTSCHJOZSA.into())], None);
-        let mut evaluator = stateful::Interpreter::new(
+        let mut evaluator = Interpreter::new(
             true,
             sources,
             PackageType::Exe,
@@ -50,7 +50,7 @@ pub fn deutsch_jozsa(c: &mut Criterion) {
 pub fn large_file(c: &mut Criterion) {
     c.bench_function("Large file parity evaluation", |b| {
         let sources = SourceMap::new([("large.qs".into(), LARGE.into())], None);
-        let mut evaluator = stateful::Interpreter::new(
+        let mut evaluator = Interpreter::new(
             true,
             sources,
             PackageType::Exe,
@@ -80,7 +80,7 @@ pub fn array_append(c: &mut Criterion) {
                 .into(),
             ),
         );
-        let mut evaluator = stateful::Interpreter::new(
+        let mut evaluator = Interpreter::new(
             true,
             sources,
             PackageType::Exe,
@@ -110,7 +110,7 @@ pub fn array_update(c: &mut Criterion) {
                 .into(),
             ),
         );
-        let mut evaluator = stateful::Interpreter::new(
+        let mut evaluator = Interpreter::new(
             true,
             sources,
             PackageType::Exe,
